test(core): relocate autodetect's inline tests into testsuite-core
ironrdp-server sets [lib] test = false, so autodetect.rs's inline
#[cfg(test)] module never ran under cargo test --workspace. Six of
its ten tests duplicate coverage already in
ironrdp-testsuite-core/tests/server/autodetect.rs, and a seventh
(sequence_number_wraps) checks the same wrap-around behavior the
existing sequence_number_wraps_at_u16_max already covers externally,
just via a private field write instead of the public API.

Move the three tests with no external equivalent
(backwards_clock_yields_a_zero_sample,
backwards_clock_keeps_the_probe_pending,
an_enormous_gap_clamps_to_u32_max) into testsuite-core and delete
the module. All ten tests' behavior is preserved: three run for the
first time under CI, seven were already redundant with the external
suite.
