+++ crates/ironrdp-testsuite-core/tests/server/autodetect.rs
@@ -96,6 +96,53 @@ fn snapshot_reflects_measurements() {
     assert_eq!(snap.avg_ms, 20);
 }
 
+/// A caller whose clock ran backwards between the request and the response must
+/// produce a zero sample. Wrapping subtraction here would yield a value near
+/// `u32::MAX` and poison every statistic drawn from the window.
+#[test]
+fn backwards_clock_yields_a_zero_sample() {
+    let mut mgr = AutoDetectManager::new();
+    let req = mgr.send_rtt_request(1000);
+
+    let response = AutoDetectResponse::RttResponse {
+        sequence_number: req.sequence_number(),
+    };
+    assert_eq!(mgr.handle_response(&response, 400), Some(0));
+
+    // The zero has to reach the window, not just the return value, since the
+    // snapshot is what the peer eventually sees.
+    let snap = mgr.snapshot().expect("one measurement was recorded");
+    assert_eq!(snap.min_ms, 0);
+    assert_eq!(snap.max_ms, 0);
+    assert_eq!(snap.avg_ms, 0);
+}
+
+/// The same backwards clock reaches expiry, where the saturation has the opposite
+/// shape: an age of zero is below any maximum, so the probe stays pending.
+/// Wrapping subtraction would make it look older than any limit and drop a probe
+/// whose response is still in flight.
+#[test]
+fn backwards_clock_keeps_the_probe_pending() {
+    let mut mgr = AutoDetectManager::new();
+    let _ = mgr.send_rtt_request(1000);
+
+    mgr.expire_stale_probes(400, 100);
+    assert_eq!(mgr.pending_count(), 1, "a probe from the future is not stale");
+}
+
+/// A gap wider than `u32::MAX` milliseconds (about 49.7 days) clamps rather than
+/// truncating to the low 32 bits, which would report a small RTT for an enormous one.
+#[test]
+fn an_enormous_gap_clamps_to_u32_max() {
+    let mut mgr = AutoDetectManager::new();
+    let req = mgr.send_rtt_request(0);
+
+    let response = AutoDetectResponse::RttResponse {
+        sequence_number: req.sequence_number(),
+    };
+    assert_eq!(mgr.handle_response(&response, u64::from(u32::MAX) + 1), Some(u32::MAX));
+}
+
 #[test]
 fn netchar_result_none_without_measurements() {
     let mut mgr = AutoDetectManager::new();
